ramfs: Lots of fixes to the build.
 * Store pointers in an addr_t instead of int32, for 64-bit's sake
 * Use DebugSupport.h instead of userlandfs Debug.h and remove extra parentheses
 * Create a header-only String class based on the userlandfs String and use it
 * RecursiveLock instead of Locker.
 * Jamfile cleanups and other misc. changes.

It isn't yet adapted to the new VFS API, so the build is still somewhat
broken.

+++ src/add-ons/kernel/file_systems/ramfs/DataContainer.cpp
@@ -4,7 +4,7 @@
 #include "Attribute.h"	// for debugging only
 #include "Block.h"
 #include "DataContainer.h"
-#include "Debug.h"
+#include "DebugSupport.h"
 #include "Misc.h"
 #include "Node.h"		// for debugging only
 #include "Volume.h"
@@ -405,7 +405,7 @@ DataContainer::_SwitchToSmallBufferMode(size_t newSize)
 	// remove the first (and only) block
 	BlockList *blocks = _GetBlockList();
 	BlockReference *block = blocks->ItemAt(0);
-	blocks->RemoveItem(0L);
+	blocks->RemoveItem((int32)0L);
 	// destroy the block list and copy the data into the small buffer
 	blocks->~BlockList();
 	if (newSize > 0)
